Remove jQuery dep and support direct links
The removal of the jQuery dependency was pretty straight-forward, since
there is only a single browser target for a Chrome extension. The one
issue was around being able to wait for elements that may asynchronously
render to a page after load events have completed. I created a helper
method just for that one purpose. That, combined with the usage of
native document APIs was sufficient to remove the jQuery dependency.

The key advantage of this removal is that it makes this package much
more acceptable for others to pick up. As it stood, security reviewers
would either have to verify the checksum of the jQuery dependency or ask
that this codebase be forked to ensure the dependency was replaced with
a known good reference. Removing the dependency side-steps the whole
concern.

Additionally, there was previously a limitation that caused this
extension to not work if a user had directly bookmarked one of the AWS
role management console links. Because the AWS SSO portal UI is skipped
in those cases, the SSO data never got saved and sent. This approach
does a save of all account names to localstorage the first time the AWS
Account SSO app is expanded in a session. From then on, the console UI
updates accordingly. The role information does not need to be saved, as
that can be retrieved from the console itself.